Enable memory tracing
This commit enables tracing of dynamic memory operations
(alloc/free/realloc) using the existing logging support in ualloc.

Note that this support is preliminary, check the code for more details.
In particular, "prevent_tracing" will have to go away eventually, but
for now we need it.

Tracing must be enabled via "yotta config":

{
    "debug": {
        "options": {
            "memory-trace": true
        }
    }
}